+++ federation/entities/activitypub/mappers.py
@@ -1,6 +1,8 @@
 from typing import List, Callable, Dict
 
 from federation.entities.activitypub.entities import ActivitypubFollow, ActivitypubProfile
+from federation.entities.base import Follow, Profile
+from federation.entities.mixins import BaseEntity
 from federation.types import UserType
 
 MAPPINGS = {
@@ -24,6 +26,45 @@ def element_to_objects(
     entities.append(cls(**transformed))
 
     return entities
+
+
+def get_outbound_entity(entity: BaseEntity, private_key: str):
+    """Get the correct outbound entity for this protocol.
+
+    We might have to look at entity values to decide the correct outbound entity.
+    If we cannot find one, we should raise as conversion cannot be guaranteed to the given protocol.
+
+    Private key of author is needed to be passed for signing the outbound entity.
+
+    :arg entity: An entity instance which can be of a base or protocol entity class.
+    :arg private_key: Private key of sender in str format
+    :returns: Protocol specific entity class instance.
+    :raises ValueError: If conversion cannot be done.
+    """
+    if getattr(entity, "outbound_doc", None):
+        # If the entity already has an outbound doc, just return the entity as is
+        return entity
+    outbound = None
+    cls = entity.__class__
+    if cls in [ActivitypubFollow, ActivitypubProfile]:
+        # Already fine
+        outbound = entity
+    elif cls == Follow:
+        outbound = ActivitypubFollow.from_base(entity)
+    elif cls == Profile:
+        outbound = ActivitypubProfile.from_base(entity)
+    if not outbound:
+        raise ValueError("Don't know how to convert this base entity to ActivityPub protocol entities.")
+    # TODO LDS signing
+    # if isinstance(outbound, DiasporaRelayableMixin) and not outbound.signature:
+    #     # Sign by author if not signed yet. We don't want to overwrite any existing signature in the case
+    #     # that this is being sent by the parent author
+    #     outbound.sign(private_key)
+    #     # If missing, also add same signature to `parent_author_signature`. This is required at the moment
+    #     # in all situations but is apparently being removed.
+    #     # TODO: remove this once Diaspora removes the extra signature
+    #     outbound.parent_signature = outbound.signature
+    return outbound
 
 
 def message_to_objects(